// src/app/Model/city.model.ts
export class CityModel {
    ilPlaka: number;
    il: string;
    ilce: string;
    enlem: number;
    boylam: number;
    merkezId: number;
    yukseklik: number;
    aciklama: string;
    alternatifHadiseIstNo: number;
    gunlukTahminIstNo: number;
    oncelik: number;
    saatlikTahminIstNo: number;
    sondurumIstNo: number;
    modelId: number;
}