// src/app/Model/district.model.ts
export class DistrictModel {
    alternatifHadiseIstNo: number;
    boylam: number;
    enlem: number;
    gunlukTahminIstNo: number;
    il: string;
    ilPlaka: number;
    ilce: string;
    merkezId: number;
    oncelik: number;
    saatlikTahminIstNo: number;
    sondurumIstNo: number;
    yukseklik: number;
    aciklama: string;
    modelId: number;
}