// src/app/Model/son.durum.model.ts
export class SonDurumModel {
    aktuelBasinc: number;
    denizSicaklik: number;
    denizeIndirgenmisBasinc: number;
    gorus: number;
    hadiseKodu: string;
    istNo: number;
    kapalilik: number;
    karYukseklik: number;
    nem: number;
    rasatMetar: string;
    rasatSinoptik: string;
    rasatTaf: string;
    ruzgarHiz: number;
    ruzgarYon: number;
    sicaklik: number;
    veriZamani: string;
    yagis00Now: number;
    yagis10Dk: number;
    yagis12Saat: number;
    yagis1Saat: number;
    yagis24Saat: number;
    yagis6Saat: number;
    denizVeriZamani: string;
}